+++ mod_app/models/__init__.py
@@ -1,0 +1,13 @@
+from .film_model import Film
+from .support_models import Actor, Analysis, Crew, Copy, Link, Tag, Location
+
+__all__ = [
+    "Actor",
+    "Analysis",
+    "Copy",
+    "Crew",
+    "Film",
+    "Link",
+    "Location",
+    "Tag",
+]
