+++ mod_app/models/film_model.py
@@ -1,111 +1,8 @@
 from django.db import models
 from django.db.models.signals import m2m_changed
 
-from ckeditor.fields import RichTextField
 
-
-class Tag(models.Model):
-    def __str__(self):
-        return self.name
-
-    name = models.CharField(max_length=255, unique=True)
-
-
-class Link(models.Model):
-    def __str__(self):
-        return self.path
-
-    path = models.URLField()
-    description = models.CharField(
-        max_length=250,
-        help_text="short description of what the link is to (optional)",
-        blank=True,
-        null=True,
-    )
-
-
-class Copy(Link):
-    class Meta:
-        verbose_name_plural = "Copies"
-
-    def __str__(self):
-        if self.name:
-            return self.name
-        else:
-            return super().__str__()
-
-    name = models.CharField(
-        max_length=100,
-        help_text="optional name/title for the copy",
-        blank=True,
-    )
-
-    CONDITION_CHOICES = [
-        ("good", "Good"),
-        ("reasonable", "Reasonable"),
-        ("poor", "Poor"),
-        ("fragile", "Fragile"),
-        ("other", "Other"),
-    ]
-
-    condition = models.CharField(
-        max_length=10,
-        choices=CONDITION_CHOICES,
-        default="reasonable",
-    )
-
-    condition_comments = models.TextField(
-        help_text="Use this to expand on the condition, particularly if 'Other'"
-    )
-
-
-class Analysis(models.Model):
-    class Meta:
-        verbose_name_plural = "Analyses"
-
-    def __str__(self):
-        return f"Analysis of {self.film}"
-
-    content = RichTextField()
-    film = models.ForeignKey(
-        "Film", on_delete=models.DO_NOTHING, related_name="analyses"
-    )
-
-
-class Actor(models.Model):
-    def __str__(self):
-        return f"{self.name}"
-
-    name = models.CharField(max_length=100)
-    birth_place = models.ForeignKey(
-        "Location", on_delete=models.SET_DEFAULT, default=None, blank=True, null=True
-    )
-    tags = models.ManyToManyField(Tag, related_name="actor_tags")
-    about = models.TextField(blank=True)
-
-
-class Crew(models.Model):
-    class Meta:
-        verbose_name_plural = "Crew Members"
-
-    def __str__(self):
-        return f"{self.name}"
-
-    ROLE_CHOICES = [
-        ("Director", "Director"),
-        ("Screenplay", "Screenplay"),
-        ("Scenography", "Scenography"),
-        ("Photography", "Photography"),
-    ]
-
-    name = models.CharField(max_length=100)
-    roles = models.CharField(max_length=255, choices=ROLE_CHOICES, blank=True)
-
-    def get_roles(self):
-        return self.roles.split(",") if self.roles else []
-
-    def set_roles(self, roles):
-        self.roles = ",".join(roles)
+from mod_app.models.support_models import Link, Tag, Copy
 
 
 class Film(models.Model):
@@ -217,19 +114,9 @@ class Film(models.Model):
     comments = models.TextField(blank=True)
 
 
-class Location(models.Model):
-    def __str__(self):
-        return f"{self.address}"
-
-    address = models.CharField(max_length=200)
-    # could use a geo package for more specific stuff, might help with google maps
-    associated_actors = models.ManyToManyField("Actor", blank=True)
-    associated_films = models.ManyToManyField("Film", blank=True)
-
-    is_setting = models.BooleanField(default=False)
-
-
 def update_actors_films(sender, instance, action, reverse, pk_set, **kwargs):
+    from mod_app.models import Actor
+
     if action == "post_add" and not reverse:
         for pk in pk_set:
             actor = Actor.objects.get(pk=pk)
