--- museum_of_dreams_project/settings/secrets_example.py
@@ -1,1 +1,0 @@
-secret_key = "not_a_secret"
